fix: allow to move legal comments from HTML
