+++ e2e/cases/html/legal-comments/rsbuild.config.ts
@@ -1,0 +1,5 @@
+export default {
+  html: {
+    template: './src/index.html',
+  },
+};
