+++ e2e/cases/html/legal-comments/src/index.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html>
+  <head></head>
+  <body>
+    <script>
+      /**
+       * @license test
+       *
+       * Copyright (c) Test
+       */
+      function test() {}
+    </script>
+  </body>
+</html>
