+++ e2e/cases/html/title/src/empty.html
@@ -1,3 +1,4 @@
-<!DOCTYPE html>
-<head></head>
+<!doctype html>
+<html>
+  <head></head>
 </html>
