+++ e2e/cases/html/title/src/plugin-options-title.html
@@ -1,5 +1,6 @@
-<!DOCTYPE html>
-<head>
-  <title><%= htmlWebpackPlugin.options.title %></title>
-</head>
+<!doctype html>
+<html>
+  <head>
+    <title><%= htmlWebpackPlugin.options.title %></title>
+  </head>
 </html>
